Recover original camera water and HUD behavior from retail

+++ src/Game/Camera/CameraManGame.cpp
@@ -19,6 +19,7 @@
 #include "Game/Util/MathUtil.hpp"
 #include "Game/Util/SceneUtil.hpp"
 #include <cstring>
+#include <stdint.h>
 
 void CameraManGame_FORCE_MATCH_SDATA2() {
     (void)1.0f;
@@ -493,7 +494,7 @@ void CameraManGame::createStartAnimCamera() {
         CameraDirector* director = CameraLocalUtil::getCameraDirector();
         chunk->setCameraType("CAM_TYPE_ANIM", director->mHolder);
 
-        chunk->mGeneralParam->mNum1 = reinterpret_cast< s32 >(data);
+        chunk->mGeneralParam->mNum1 = reinterpret_cast< intptr_t >(data);
         chunk->_64 = true;
     }
 }
